Radically simplify the project.
This thing had really gotten out of hand. There was
a ton of code that was experimental and poorly thought
out that had made it into master. This commit pares
everything down.

As of this commit, this app finds PDF’s on the NYPD
website, downloads them, and re-uploads them to an
S3 account.